requirements.txt
asgiref==3.8.1
Django==5.1.7
gunicorn==23.0.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.1
whitenoise==6.9.0
django-widget-tweaks